+++ my-app/src/App.css
@@ -1,4 +1,4 @@
 #root {
-    max-width: 1280px;
-    margin: 0 auto;
-  }
+  max-width: 1280px;
+  margin: 0 auto;
+}
