+++ my-app/src/Components/Button.tsx
@@ -1,7 +1,7 @@
 function Button({ text, onCustomClick, styles }: {text: string, onCustomClick?: Function, styles?: string }) {
-    return (
-        <button className={styles} onClick={() => onCustomClick?.()}>{text}</button>
-    )
+  return (
+    <button className={styles} onClick={() => onCustomClick?.()}>{text}</button>
+  )
 }
 
 export default Button
