+++ my-app/src/Components/List.tsx
@@ -1,7 +1,7 @@
 function List({ customList }: {customList: string[]}) {
-    return (
-        <ul className="text-center">{customList.map(n => <li key={n}>{n}</li>)}</ul>
-    )
+  return (
+    <ul className="text-center">{customList.map(n => <li key={n}>{n}</li>)}</ul>
+  )
 }
 
 export default List
