+++ my-app/src/Components/index.tsx
@@ -2,6 +2,6 @@ import Button from './Button'
 import List from './List'
 
 export {
-    Button,
-    List
+	Button,
+	List
 }
