+++ my-app/src/Layout/index.tsx
@@ -1,0 +1,7 @@
+import NavBar from './NavBar'
+import Body from './Body'
+
+export {
+	NavBar,
+	Body,
+}
